Ajuste del tipo de dato de las cantidades de rcd y ajuste del bug de editar

+++ sistema-certificados-backend/src/main/java/com/agg/certificados/dtos/request/QuantitiesRcdRequestDto.java
@@ -2,14 +2,14 @@ package com.agg.certificados.dtos.request;
 
 
 public class QuantitiesRcdRequestDto {
-    public Long quantity_rcd_1;
-    public Long quantity_rcd_2;
-    public Long quantity_rcd_3;
-    public Long quantity_rcd_4;
-    public Long quantity_rcd_5;
-    public Long quantity_rcd_6;
-    public Long quantity_rcd_7;
-    public Long quantity_rcd_8;
-    public Long quantity_rcd_9;
+    public Double quantity_rcd_1;
+    public Double quantity_rcd_2;
+    public Double quantity_rcd_3;
+    public Double quantity_rcd_4;
+    public Double quantity_rcd_5;
+    public Double quantity_rcd_6;
+    public Double quantity_rcd_7;
+    public Double quantity_rcd_8;
+    public Double quantity_rcd_9;
 
 }
