+++ sistema-certificados-backend/src/main/java/com/agg/certificados/dtos/request/QuantitiesTotal.java
@@ -2,5 +2,5 @@ package com.agg.certificados.dtos.request;
 
 public class QuantitiesTotal {
     public QuantitiesRcdRequestDto quantities;
-    public Long total;
+    public double total;
 }
