+++ sistema-certificados-backend/src/main/java/com/agg/certificados/dtos/response/ReportCvcResponseDto.java
@@ -5,9 +5,9 @@ public class ReportCvcResponseDto {
     public String number_id;
     public String legal_representative;
     public String activities_by_manager;
-    public Long t_puntolimpio;
-    public Long t_aprovechamiento;
-    public Long t_dispuesto_finalmente;
+    public double t_puntolimpio;
+    public double t_aprovechamiento;
+    public double t_dispuesto_finalmente;
 
 
 }
